Rename to request identity
